Implemented Scriptable Objects for Audio
Small, and albeit unnecessary change. Used ScriptableObjects to store large audio clip arrays. I had already done this the hard way so there's no real gain for this project, but the PlayClip script that works off of scriptables should be helpful for future projects

essentially it's the same, but now the aduio clips are stored on a scriptableobject. This means say if you want to have  10 objects play from the same list of 5 random audio clips, instead of having to manually give each one those 5 clips for a total of performing this action 50 times. Now you you just give each the scriptableobject, reducing manual asignments to 10 actions instead.